refactor(#162): extract sheet.py — compose-then-pack scale/page selection (P3b)
Move the outer-layout cluster (ADR 0004) out of make_drawing.py into sheet.py:
the strip-depth estimators, the StripDepths/AnnoBox/ViewBlock footprint model,
_measure_strips/_compose_anno_boxes/_footprint_from_boxes/_fits, choose_scale, and
_layout_geometry/_build_zones (+ _padded_box, _view_geom, _attribute_annotations,
the _DIM_PAD invariant assert). Builds on P3a, which put the shared constants in
_core.

sheet.py sits below make_drawing: it imports only _core + build123d_drafting, and
make_drawing re-imports the eight symbols its staying code (build flow, _assemble,
_repack) still calls — all downward, no cycle. The measure-and-repack pass
(_repack/_repack_candidates and its overlap/measure helpers) stays in make_drawing
because it re-runs _assemble (orchestration → P6).

Tests that imported these internals from make_drawing now import them from their
real homes (sheet.py / _core). Pure move: golden gate passes against the committed
snapshots with no regeneration; the full make_drawing/golden/layout suites pass
(389); ruff check + format + mypy clean. Adversarially checked: sheet imports
nothing upward (no cycle, 0 callbacks), no stale defs, the _DIM_PAD assert still
runs at import, and box/cylinder/holed parts build views lint-clean.

Closes #162.
