+++ docs/plans/138-module-split-roadmap.md
@@ -34,7 +34,7 @@ no PR introduces an import cycle, riskiest (annotations) last:
 |---|---|---|---|---|
 | ~~P1~~ | #160 | `_text_width` → `_core` ✅ | tiny | — |
 | ~~P2~~ | #161 | `projection.py` (silhouettes, iso) ✅ | med | — |
-| **P3** | #162 | `sheet.py` — 3a footprints/estimators, 3b scale/zone/repack | large (2 PRs) | P1 |
+| ~~P3~~ | #162 | `sheet.py` — compose-then-pack (repack deferred to P6) ✅ | large (2 PRs) | P1 |
 | **P4** | #163 | `analysis.py` (`_analyse`) | med-lg | — |
 | **P5** | #164 | `annotations/` — sections, turned, pmi, **holes**, **orchestrator** | biggest (~5 PRs) | P1–P4 |
 | **P6** | #165 | `builder.py` + thread the build context | med | P2, P4 |
@@ -45,7 +45,7 @@ no PR introduces an import cycle, riskiest (annotations) last:
 make_drawing.py   # transitional compat facade / public re-exports
 builder.py        # build_drawing/make_drawing orchestration (P6)
 analysis.py       # _analyse, Analysis construction (P4)
-sheet.py          # choose_scale, compose-then-pack, repack (P3)
+sheet.py          # choose_scale, compose-then-pack (DONE; repack→P6)
 projection.py     # view projection, silhouettes, iso fit (DONE)
 registry.py       # annotation identity (DONE)
 linting.py        # lint_feature_coverage, _suggest_fix, CoverageState (DONE)
